+++ CHANGELOG.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Changed
+- Added Skills Self-Evaluation entry in Academics section in Browse panel [#2746](https://github.com/rokwire/illinois-app/issues/2746).
+### Changed
 - Updated GroupAdvancedSettingsPanel [#2744](https://github.com/rokwire/illinois-app/issues/2744).
 - Updated privacy content statements [#2763](https://github.com/rokwire/illinois-app/issues/2763).
 ### Fixed
 - Fixed privacy content statement strings [#2763](https://github.com/rokwire/illinois-app/issues/2763).
+- Unread notifications issues [#2761](https://github.com/rokwire/illinois-app/issues/2761).
+- Cleaned up Create Group / Research Project button processing [#2740](https://github.com/rokwire/illinois-app/issues/2740).
 
 ## [4.2.41] - 2022-12-14
 ### Fixed